refactor(ai): dedupe Tool Inventory logic into tool_inventory_materials
